+++ src/constants/Extension.ts
@@ -6,6 +6,8 @@ export const CONFIG_INCLUDE_FOLDERS = 'config.includeFoldersToExport';
 export const CONFIG_SEMIS = 'config.semis';
 export const CONFIG_QUOTE = 'config.quote';
 
+export const EXTENSION_NAME = "Barrel Generator";
+
 export const COMMAND_KEYS = {
   Generate: "generate",
   Open: "open",